Fixed cookie.js error
If we are going to change html via cookie reader function, we have to load the html before we try to change it. I moved the cookie read function, but it seems counterintuitive, because the FIRST thing we should do when loading the page is checking the cookie. Summary: my solution is a little hacky